Fixes install script

+++ install.sh
@@ -17,9 +17,11 @@ echo -n "Changing to the $dir directory ..."
 cd $dir
 echo "done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
+echo "Moving any existing dotfiles from ~ to $olddir"
 for file in $files; do
-	echo "Moving any existing dotfiles from ~ to $olddir"
-	mv ~/$file ~/dotfiles_old/
+	if [[ -e ~/$file ]] ; then
+		mv ~/$file ~/dotfiles_old/
+	fi
 	echo "Creating symlink to $file in home directory."
 	ln -s $dir/$file ~/$file
 done
@@ -27,8 +29,8 @@ install_zsh () {
 	# Test to see if zshell is installed. If it is:
 	if [ -f /bin/zsh -o -f /usr/bin/zsh ]; then
 		# Clone my oh-my-zsh repository from GitHub only if it isn't already present
-		if [[ ! -d $dir/oh-my-zsh/ ]]; then
-			git clone http://github.com/robbyrussell/oh-my-zsh.git
+		if [[ ! -d $dir/.oh-my-zsh/ ]]; then
+			git clone http://github.com/robbyrussell/oh-my-zsh.git .oh-my-zsh
 		fi
 		# Set the default shell to zsh if it isn't currently set to zsh
 		if [[ ! $(echo $SHELL) == $(which zsh) ]]; then
